optimization/autoagent/pyproject.toml
[project]
name = "indiios-autoagent"
version = "0.1.0"
description = "System prompt optimization for indiiOS agents"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.73.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.2",
    "requests>=2.31.0",
]

[tool.uv]
managed = true
